+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `Signal#prepend` method
 * Add `Signal#append` method
 * Fix an issue where exceptions were being swallowed when mounting
+* Add `takeUntil` function
 
 ## 1.4.0 (2019-01-08)